--- morib/src/main/java/org/morib/server/api/homeViewApi/service/fetch/timer/FetchTimerService.java
@@ -1,7 +1,0 @@
-package org.morib.server.api.homeViewApi.service.fetch.timer;
-
-import java.time.LocalDate;
-
-public interface FetchTimerService {
-    void fetch();
-}
