--- morib/src/main/java/org/morib/server/api/homeViewApi/service/fetch/timer/FetchTimerServiceImpl.java
@@ -1,16 +1,0 @@
-package org.morib.server.api.homeViewApi.service.fetch.timer;
-
-import lombok.RequiredArgsConstructor;
-import org.morib.server.domain.timer.application.TimerGateway;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class FetchTimerServiceImpl implements FetchTimerService {
-    private final TimerGateway timerGateway;
-
-    @Override
-    public void fetch() {
-
-    }
-}
